+++ vsd/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `save`
+  - Support for fake png header segments.
+
 ## [0.4.3] - 2025-08-16
 
 ### Added